LPS-60317 Move tuple class to toplevel
